monday-learn-api/requirements.txt
fastapi
uvicorn[standard]
loguru
sqlalchemy
pymysql
python-dotenv
pydantic-settings
bcrypt
email-validator
python-jose[cryptography]
python-multipart
httpx
